+++ core/src/main/java/com/learning/api/exception/JavaTestException.java
@@ -1,4 +1,4 @@
-package com.learning.kotlinapi.exception;
+package com.learning.api.exception;
 
 public class JavaTestException extends Exception {
   public JavaTestException(String message) {
