+++ core/src/main/kotlin/com/learning/api/config/Logger.kt
@@ -1,4 +1,4 @@
-package com.learning.kotlinapi.config
+package com.learning.api.config
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
