+++ core/src/main/kotlin/com/learning/api/exception/KotlinTestException.kt
@@ -1,3 +1,3 @@
-package com.learning.kotlinapi.exception
+package com.learning.api.exception
 
 class KotlinTestException(override val message: String) : Exception()
